refactor: split campaign dashboard sections
